Undo rename: `_figure_utils.py` -> `_matplotlib`
