+++ src/pymovements/plotting/scanpathplot.py
@@ -30,10 +30,10 @@ from matplotlib.patches import Circle
 
 from pymovements.events import EventDataFrame
 from pymovements.gaze import Gaze
-from pymovements.plotting._figure_utils import _draw_line_data
-from pymovements.plotting._figure_utils import _setup_axes_and_colormap
-from pymovements.plotting._figure_utils import finalize_figure
-from pymovements.plotting._figure_utils import LinearSegmentedColormapType
+from pymovements.plotting._matplotlib import _draw_line_data
+from pymovements.plotting._matplotlib import _setup_axes_and_colormap
+from pymovements.plotting._matplotlib import finalize_figure
+from pymovements.plotting._matplotlib import LinearSegmentedColormapType
 
 # This is really a dirty workaround to use the Agg backend if runnning pytest.
 # This is needed as Windows workers on GitHub fail randomly with other backends.
