+++ src/pymovements/plotting/traceplot.py
@@ -27,10 +27,10 @@ import matplotlib.scale
 import numpy as np
 
 from pymovements.gaze.gaze import Gaze
-from pymovements.plotting._figure_utils import _draw_line_data
-from pymovements.plotting._figure_utils import _setup_axes_and_colormap
-from pymovements.plotting._figure_utils import finalize_figure
-from pymovements.plotting._figure_utils import LinearSegmentedColormapType
+from pymovements.plotting._matplotlib import _draw_line_data
+from pymovements.plotting._matplotlib import _setup_axes_and_colormap
+from pymovements.plotting._matplotlib import finalize_figure
+from pymovements.plotting._matplotlib import LinearSegmentedColormapType
 
 # This is really a dirty workaround to use the Agg backend if runnning pytest.
 # This is needed as Windows workers on GitHub fail randomly with other backends.
